Move Format classes into separate module
